==> hackathon-nt-2024/datasets/data/Billing/02 read_billing_data.ipynb <==
# วิเคราะห์และเตรียมข้อมูลจาก Billing files

## Library ที่จำเป็นต้องใช้ 

import pandas as pd
from scipy import stats

## Function แปลงข้อความแบบวันที่ ให้เป็น dataframe date

# Function to convert BE to AD and parse the date
def convert_be_to_ad(date_int):
    if pd.isnull(date_int):
        return None
    date_str = str(int(date_int)).zfill(8)  # Ensure the string is 8 characters long
    day = int(date_str[:2])
    month = int(date_str[2:4])
    year_be = int(date_str[4:])
    year_ad = year_be - 543
    return pd.Timestamp(year_ad, month, day)

## อ่านไฟล์ข้อมูล Billing ที่รวมไฟล์มาแล้ว

billing_df = pd.read_csv("BP_2023_V1.csv", on_bad_lines="warn", dtype=str)

## ดูจำนวน column

len(billing_df.columns)


## แสดงชื่อ column

billing_df.columns

# สำรวจข้อมูล 

billing_df

## ดูข้อมูลใน column เพื่อดูว่าประกอบด้วยข้อมูลอะไรบ้าง

data_frequencies = billing_df["ACCOUNT_TYPE"].value_counts().sort_values(ascending=False)
print("\nData Frequencies:")
print(data_frequencies)

data_frequencies = billing_df["LD CHARGE"].value_counts().sort_values(ascending=False)
print("\nData Frequencies:")
print(data_frequencies)

billing_df.head()

##  มีข้อมูลที่ไม่จำเป็นหรือไม่ หรือข้อมูลที่เป็นค่าว่าง หรือไม่มีข้อมูลอยู่

billing_df.isnull().sum()


นับจำนวนบรรทัด

len(billing_df)

billing_df.isna().sum().sort_values()

## เลือก column ที่จำเป็นต้องใช้

billing_df = billing_df[[ 
       'INACTIVE_DATE', 'ACCOUNT_TYPE', 'TEL_NO',
       'ACTIVE_DATE', 'ACCOUNT_NO', 
       'BANK_CODE', 'CUSTOMER_REF', 'AMOUNT_DUE',
       'SERVICE_LOCATION NAME', 'PRODUCT DESC',
       'BILL MONTH']]
billing_df

billing_df.columns

## แปลง DATE Columns ให้เป็นรูปแบบวันที่ เช่น ACTIVE_DATE, INACTIVE_DATE

# Apply the conversion function to the column
# Convert INACTIVE_DATE to integers, ignoring errors
billing_df['INACTIVE_DATE_INT'] = pd.to_numeric(billing_df['INACTIVE_DATE'], errors='coerce')

# Apply the function to the new integer column
billing_df['INACTIVE_DATE'] = billing_df['INACTIVE_DATE_INT'].apply(convert_be_to_ad)

# Drop the intermediate integer column
billing_df.drop(columns=['INACTIVE_DATE_INT'], inplace=True)

billing_df['INACTIVE_DATE']

# Filter rows where INACTIVE_DATE is not NaT
valid_dates_df = billing_df[billing_df['INACTIVE_DATE'].notna()]
valid_dates_df

# Apply the conversion function to the column
# Convert INACTIVE_DATE to integers, ignoring errors
billing_df['ACTIVE_DATE_INT'] = pd.to_numeric(billing_df['ACTIVE_DATE'], errors='coerce')

# Apply the function to the new integer column
billing_df['ACTIVE_DATE'] = billing_df['ACTIVE_DATE_INT'].apply(convert_be_to_ad)

# Drop the intermediate integer column
billing_df.drop(columns=['ACTIVE_DATE_INT'], inplace=True)

# Filter rows where ACTIVE_DATE is NaT
valid_dates_df = billing_df[billing_df['ACTIVE_DATE'].isna()]
valid_dates_df

valid_dates_df = billing_df[billing_df['ACTIVE_DATE'].notna()]
date_frequencies = valid_dates_df['ACTIVE_DATE'].value_counts().sort_values(ascending=False)
print("\nDate Frequencies:")
print(date_frequencies)

data_frequencies = billing_df["ACCOUNT_TYPE"].value_counts().sort_values(ascending=False)
print("\nData Frequencies:")
print(data_frequencies)
data_frequencies.to_csv("ACCOUNT_TYPE.csv")

data_frequencies = billing_df["PRODUCT DESC"].value_counts().sort_values(ascending=False)
print("\nData Frequencies:")
print(data_frequencies)

data_frequencies.to_csv("PRODUCT.csv")

# Group by ACCOUNT_NO and aggregate AMOUNT_DUE into a list
grouped_df = billing_df.groupby('ACCOUNT_NO')['AMOUNT_DUE'].apply(list).reset_index()
grouped_df
grouped_df.to_csv("ACCOUNT_NO_GROUP.csv")

## หาค่าเฉลี่ยของ bill

billing_df.loc[:,"AMOUNT_DUE"] = pd.to_numeric(billing_df["AMOUNT_DUE"], errors='coerce')

# Group DataFrame by 'ACCOUNT_NO' and apply the summary function
# summary_df = billing_df.groupby('ACCOUNT_NO').apply(summarize_by_account)

# Group DataFrame by 'ACCOUNT_NO' and calculate average
average_due = billing_df.groupby('ACCOUNT_NO')['AMOUNT_DUE'].mean()


average_due

billing_df.columns

billing_df.loc[:,"AMOUNT_DUE"] = pd.to_numeric(billing_df["AMOUNT_DUE"], errors='coerce')

summary_df = billing_df.groupby(['ACCOUNT_NO', 'TEL_NO', 'PRODUCT DESC', 'ACCOUNT_TYPE', 
                                  'BANK_CODE', 'CUSTOMER_REF', 'SERVICE_LOCATION NAME',
                                  'INACTIVE_DATE', 'ACTIVE_DATE'])['AMOUNT_DUE'].mean()

summary_df


summary_df.to_csv("SUMMARY_BILLING_2023.csv")
