==> hackathon-nt-2024/datasets/data/Billing/03 read_summary_billing.ipynb <==
import pandas as pd

pd.options.display.float_format = '{:,.2f}'.format

summary_billing_df = pd.read_csv("SUMMARY_BILLING_2023.csv")

summary_billing_df.dtypes

summary_billing_df.head()

# Group by 'PRODUCT DESC' and sum 'AMOUNT_DUE'
product_due_sum = summary_billing_df.groupby('PRODUCT DESC')['AMOUNT_DUE'].sum().sort_values(ascending=False)
product_due_sum

# List of desired product descriptions
desired_products = ['Fiber 2U', 'IP Phone']
# Filter rows based on 'PRODUCT DESC' using boolean indexing
filtered_df = summary_billing_df[summary_billing_df['PRODUCT DESC'].isin(desired_products)]

filtered_df

filtered_df.groupby(['ACCOUNT_TYPE', 'PRODUCT DESC'])['AMOUNT_DUE'].sum().sort_values(ascending=False)

filtered_df.to_csv("Customer Group and Product Amount.csv", index=False, float_format="%.2f")

filtered_df.groupby(['ACCOUNT_TYPE'])['AMOUNT_DUE'].sum().sort_values(ascending=False)

filtered_df[['ACCOUNT_TYPE', 'PRODUCT DESC']].value_counts().sort_values(ascending=False)
