==> hackathon-nt-2024/datasets/data/Billing/04 billing_product_categories.ipynb <==
# การจัดกลุ่ม Package ของลูกค้า Residential
## เพื่อกำหนดกลุ่มเป้าหมายในการจัดการ

import pandas as pd
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.cluster import KMeans
import matplotlib.pyplot as plt
from matplotlib.font_manager import fontManager

pd.options.display.float_format = '{:,.2f}'.format
# ปิดการแสดงคำเตือน SettingWithCopyWarning ชั่วคราว
pd.options.mode.chained_assignment = None  # ปิดการแสดงคำเตือน


# import matplotlib as mpl

# Add font to font manager
fontManager.addfont('/Users/seal/Documents/GitHub/hackathon-nt-2024/fonts/Sarabun/Sarabun-Regular.ttf')
# Set the font for your plot
plt.rcParams['font.family'] = 'Sarabun'
# Set font size for all plot elements
plt.rcParams['font.size'] = 12
# mpl.rc("font", family="Sarabun", size=24)

## อ่านข้อมูลจาก billing ซึ่งมีข้อมูล package ที่ลูกค้าใช้อยู่

df = pd.read_csv("CONTRACT_PACKAGE_NAME_20240605.csv")

## ดูประเภทลูกค้า

# Assuming you have a DataFrame called df
unique_values = df['ACCOUNT_TYPE'].unique()

print(unique_values)

## เลือกเฉพาะลูกค้า Residential

df_target = df[df['ACCOUNT_TYPE'] == 'Residential - Normal']
df_target

### จัดการข้อมูลที่ไม่ได้ระบุประเภท package

df_target['CONTRACRT NAME TH'] = df_target['CONTRACRT NAME TH'].fillna("")

# ทำการแบ่งหมวดหมู่ package
1. Preprocess the text data: Convert the text in the "CONTRACRT NAME TH" column into numerical vectors using techniques like TF-IDF.
2. Apply a clustering algorithm: Use K-means or another clustering algorithm to cluster the data into different categories.
3. Assign cluster labels: Use the resulting cluster labels as the new "CATEGORY" column.


# Convert the text data into numerical vectors using TF-IDF
tfidf_vectorizer = TfidfVectorizer()
tfidf_matrix = tfidf_vectorizer.fit_transform(df_target['CONTRACRT NAME TH'])

# Apply K-means clustering
num_clusters = 15  # Adjust the number of clusters as needed, ปรับตัวเลขตามต้องการ ความเหมาะสมอยู่ที่ทดสอบแล้วดูผลลัพธ์
kmeans = KMeans(n_clusters=num_clusters, random_state=42)
kmeans.fit(tfidf_matrix)

# Assign cluster labels as the new "CATEGORY" column
df_target['CATEGORY'] = kmeans.labels_ + 1  # Adding 1 to start categories from 1

# Display the resulting DataFrame
df_target

# Group by CATEGORY
grouped_df = df_target.groupby('CATEGORY').apply(lambda x: x).reset_index(drop=True)
grouped_df

# นำผลลัพธ์ไปใช้งาน

grouped_df.to_excel("/Users/seal/Documents/GitHub/hackathon-nt-2024/datasets/data/Billing/PACKAGE_CATEGORY_20240605.xlsx", index=False)
